Updated Payload and Response names
As we begin the process of making significant changes to the structure and functionality methods we needed to organize the objects we have.
This commit removes an unneeded object, and updates the javadocs of the others.
In addition the observationNotes end point was removed as that end point will no longer be handled on this RESTful client.