Added Meeseeks Box (Thread Pooling)
Added the needed objects to allow for thread pooling to handle the act of taking the data from each sensor and putting it into the database accordingly.

This is the beginning of a long process to make the client faster and to setup the application to put the data into numorous tables based on the different time frames a user may query the database for data.