Fixed Thread Issue
This is a patch that fixes threads doing the same work as each other by adding synchronization. Some object names were updated as well because the work that the threads do has been updated.
It is known that the persistant data currently does not keep track of timing but knows how to handle higher and lower values based on the database structure.